+++ im_model_gen/test/gen_test_cases/index.dart
@@ -9,7 +9,6 @@ part 'test_case_inheritance.dart';
 part 'test_case_inheritance_abstract.dart';
 part 'test_case_inverted_flags.dart';
 part 'test_case_no_copy.dart';
-part 'test_case_object.dart';
 part 'test_case_primary_constructor.dart';
 part 'test_case_private_named_parameters.dart';
 part 'test_case_static_const.dart';